Update file Loey, correzione errori

+++ src/main/java/com/gestionePersonale/demo/Dao/OrarioLavoroDao.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface OrarioLavoroDao extends JpaRepository<OrarioLavoro, Integer> {
-    List<OrarioLavoro> findByPersonaleId(Long personaleId);
+    List<OrarioLavoro> findByPersonaleId(Integer personaleId);
 }
